\name{PiMP-package}
\alias{PiMP-package}
\alias{PiMP}
\docType{package}
\title{
What the package does (short line)
PiMP provides functions to process and analyse mzXML and mzML datasets.
}
\description{
PiMP is a collection of functions that can be used to process and analyse untargetted LCMS data. It is based around the mzmatch.R package.  Datasets can be processed using a single function, Pimp.runPipeline, generating an Microsoft Excel report.  It is also used by the PiMP webtool.
}
\details{
\tabular{ll}{
Package: \tab PiMP\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-08-19\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
Pimp.runPipeline(files=files, groups=groups, standards=stds, contrasts=contrasts, databases=databases, nSlaves=4, reports="excel")
}
\author{
Moi

Maintainer: Who to complain to <gp-software@glasgow.ac.uk>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
